Separation of concerns between api endpoint, simulation and impact result and use base-amendement scenario names
